requirements.txt
boto3==1.34.69
PyYAML==6.0.1
requests==2.31.0
